refactor: :recycle: refactored styling code and custom components code.
