+++ src/Components/NutritionStats/types.ts
@@ -1,0 +1,8 @@
+export type NutritionStatsProps = {
+    item: {
+      title: string;
+      quantity: string | number;
+      percentage: string | number;
+      color: string;
+    };
+  };
