+++ src/Components/PieChartInfoItem/types.ts
@@ -1,0 +1,5 @@
+export type PieChartInfoItemProps = {
+  item: string;
+  percentage: string | number;
+  color: string;
+};
